#2 Part 2 done VIEWER-8

+++ src/include/matrix_calc.h
@@ -15,9 +15,9 @@ enum { X_AXIS, Y_AXIS, Z_AXIS, N_AXES };
  * @author yukikoqo
  */
 typedef struct {
-  double rotation_angles[N_AXES];
-  double translation_vector[N_AXES];
-  double scale_vector[N_AXES];
+  float rotation_angles[N_AXES];
+  float translation_vector[N_AXES];
+  float scale_vector[N_AXES];
 } mvp_matrix_movement_t;
 
 /** @brief Матрица преобразования
